+++ app/models/article.rb
@@ -3,4 +3,8 @@ class Article < ApplicationRecord
 
   validates :title, presence: true
   validates :content, presence: true
+
+  def draft?
+    !published?
+  end
 end
