+++ app/policies/articles/publication_policy.rb
@@ -1,0 +1,9 @@
+module Articles
+  class PublicationPolicy < ApplicationPolicy
+    alias article record
+
+    def create?
+      user&.administrative_role?
+    end
+  end
+end
